// packages/api-client/src/generated/models/LocateSearchByPosition.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Coordinate } from './Coordinate';
import type { LocateOptions } from './LocateOptions';
export type LocateSearchByPosition = {
    coordinate?: Coordinate;
    options?: LocateOptions;
};